Fix an editor bug that prevented the editor from being styled when the file was small.
